Moved code directly to view
