+++ src/senaite/storage/browser/__init__.py
@@ -19,15 +19,10 @@
 # Some rights reserved, see README and LICENSE.
 
 import collections
-import json
 
 from bika.lims import api
 from Products.Five.browser import BrowserView
 from senaite.storage import logger
-from senaite.storage import senaiteMessageFactory as _
-from senaite.storage.catalog import STORAGE_CATALOG
-
-DISPLAY_TEMPLATE = "<a href='${url}' _target='blank'>${get_full_title}</a>"
 
 
 class BaseView(BrowserView):
@@ -71,50 +66,3 @@ class BaseView(BrowserView):
         """Set a portal status message
         """
         return self.context.plone_utils.addPortalMessage(message, level)
-
-    def get_reference_widget_attributes(self, name, obj):
-        """Return input widget attributes for the ReactJS component
-        """
-        url = api.get_url(obj)
-
-        attributes = {
-            "data-name": name,
-            "data-values": [],
-            "data-records": {},
-            "data-value_key": "uid",
-            "data-value_query_index": "UID",
-            "data-api_url": "%s/referencewidget_search" % url,
-            "data-query": {
-                "portal_type": ["StorageSamplesContainer"],
-                "is_full": False,
-                "review_state": "active",
-                "sort_on": "getId",
-                "sort_order": "ascending",
-            },
-            "data-catalog": STORAGE_CATALOG,
-            "data-search_index": "listing_searchable_text",
-            "data-search_wildcard": True,
-            "data-allow_user_value": False,
-            "data-columns": [{
-                "name": "id",
-                "label": _("Id"),
-                "width": 10,
-            }, {
-                "name": "get_full_title",
-                "label": _("Container path"),
-                "width": 90,
-            }],
-            "data-display_template": DISPLAY_TEMPLATE,
-            "data-limit": 5,
-            "data-multi_valued": False,
-            "data-disabled": False,
-            "data-readonly": False,
-            "data-required": False,
-            "data-clear_results_after_select": False,
-        }
-
-        for key, value in attributes.items():
-            # convert all attributes to JSON
-            attributes[key] = json.dumps(value)
-
-        return attributes
